Add Account model for control balance

+++ PersonalBankServices/Services/TransferService.cs
@@ -11,12 +11,14 @@ namespace PersonalBankServices.Repositories
     {
         //TODO implement try catch
         private ITransferRepository _repository;
+        private IAccountRepository _accountRepository;
         private IMapper _mapper;
 
-        public TransferService(ITransferRepository repository, IMapper mapper)
+        public TransferService(ITransferRepository repository, IMapper mapper, IAccountRepository accountRepository)
         {
             _repository = repository;
             _mapper = mapper;
+            _accountRepository = accountRepository;
         }
 
         public async Task<List<ReadTransferDto>> GetAllTransfers()
@@ -35,10 +37,21 @@ namespace PersonalBankServices.Repositories
 
         public async Task<ReadTransferDto> AddTransfer(CreateTransferDto transferDto)
         {
-            TransferModel transfer = _mapper.Map<TransferModel>(transferDto);
-            await _repository.AddTransfer(transfer);
+            AccountModel balance = await _accountRepository.GetBalance();
+            if (transferDto.Amount > 0 && balance.ActualBalance > transferDto.Amount)
+            {
+                TransferModel transfer = _mapper.Map<TransferModel>(transferDto);
+                balance.ActualBalance -= transfer.Amount;
+                if (balance.ActualBalance < 0)
+                {
+                    balance.ActualBalance = 0;
+                }
+                await _accountRepository.UpdateBalance(balance);
+                await _repository.AddTransfer(transfer);
 
-            return _mapper.Map<ReadTransferDto>(transfer);
+                return _mapper.Map<ReadTransferDto>(transfer);
+            }
+            throw new Exception($"Transfer wasn't suceeded");
         }
 
         public async Task<ReadTransferDto> UpdateTransfer(UpdateTransferDto transferDto)
@@ -47,23 +60,40 @@ namespace PersonalBankServices.Repositories
 
             if (transferFounded != null)
             {
+                var balance = await _accountRepository.GetBalance();
+                balance.ActualBalance += transferFounded.Amount;
+
                 var transferMapped = _mapper.Map<TransferModel>(transferDto);
+                balance.ActualBalance -= transferMapped.Amount;
+                if (balance.ActualBalance < 0)
+                {
+                    balance.ActualBalance = 0;
+                }
+
+                await _accountRepository.UpdateBalance(balance);
 
                 var transferChanged = await _repository.UpdateTransfer(transferMapped);
-
                 return _mapper.Map<ReadTransferDto>(transferChanged);
+     
             }
 
-            throw new Exception($"Transfer for id: {transferDto.Id} wasn't found");
+            throw new Exception($"Transfer wasn't suceeded");
         }
 
         public async Task<bool> DeleteTransfer(int id)
         {
-            bool transferDeleted = await _repository.DeleteTransfer(id);
+            var transferFounded = await SearchById(id);
 
-            if (transferDeleted != false)
+            if (transferFounded != null)
             {
-                return true;//TODO sucess message
+                var balance = await _accountRepository.GetBalance();
+                balance.ActualBalance += transferFounded.Amount;
+                if (balance.ActualBalance < 0)
+                {
+                    balance.ActualBalance = 0;
+                }
+                await _accountRepository.UpdateBalance(balance);
+                return await _repository.DeleteTransfer(id);
             }
 
             throw new Exception($"Transfer for id: {id} wasn't found");
